Crate lift_hom

Given an element in $\Ext(M, N)$, this computes the induced map $\Ext(N, k) \to \Ext(M, k)$ given by composition.

It begins by asking for the two modules $M$, $N$ and the $\Ext$ class. Afterwards, you may either supply save files for the two modules, or a range to compute the map for.

Afterwards, the user is prompted for the Ext class. If $R_s$ is the $s$th term of the minimal resolution of $M$, the Ext class is given as an element of $\Hom_A(R_s, \Sigma^t N) = \Hom(\Ext^{s, *}(M, k)^\vee, \Sigma^t N)$.

In other words, for every basis element in $\Ext^{s, *}(M, k)$, one has to specify its image in $\Sigma^t N$. In the special case where $s = 0$, this is specifying the map between the underlying modules on module generators under the Steenrod action.

Our notation is as follows:

f is the map in $\Hom_A(R_s, \Sigma^t N)$.

F is the induced map on Ext.

Each prompt will be of the form f(x_(s, n, i)) = and the user has to input the value of the homomorphism on this basis element. For example, the following session computes the map induced by the projection of spectra $C2 \to S^1$

 $ cargo run --example lift_hom
Target module (default: S_2): C2
Source module (default: Cnu): S_2
s of Ext class (default: 0): 0
n of Ext class (default: 0): -1
Target save file (optional):
Max target s (default: 10): 10
Max target n (default: 10): 20

Input module homomorphism to lift:
f(x_(0, 0, 0)): [1]

It is important to keep track of varaince when using this example; Both $\Ext(-, k)$ and $H^*(-)$ are contravariant functors. The words “source” and “target” refer to the map between Steenrod modules.